feat: added basic sub link
